@@ -1,12 +1,9 @@
-import './index.css';
-import './print.css';
-
-import React from 'react';
-
 import { BasicInfo } from '../BasicInfo';
 import { Education } from '../Education';
 import { Experiences } from '../Experiences';
 import { Skills } from '../Skills';
+import './index.css';
+import './print.css';
 
 export default () => (
   <div className="hero">
