@@ -1,36 +1,47 @@
-import './Experiences.css';
-
-import React from 'react';
-
-import { SectionHeader } from '../SectionHeader';
-import EmmaLogo from './emma-logo.jpg';
-import HyperLogo from './hyper-logo.jpg';
-import UsherLogo from './usher-logo.jpg';
-import WorkstationLogo from './workstation-logo.jpg';
+import { SectionHeader } from "../SectionHeader";
+import EmmaLogo from "./emma-logo.jpg";
+import HyperLogo from "./hyper-logo.jpg";
+import LightingLogo from "./lightning-logo.png";
+import OrangeLogo from "./orange-logo.png";
+import UsherLogo from "./usher-logo.jpg";
+import WorkstationLogo from "./workstation-logo.jpg";
 
 export const Experiences = () => (
   <section>
     <SectionHeader title="Experiences" />
     <div className="columns is-2">
       <div className="column is-4">
-        <a
-          target="_blank"
-          rel="noopener noreferrer"
-          href="https://www.microstrategy.com/"
-        >
+        <a target="_blank" rel="noopener noreferrer" href="https://www.microstrategy.com/">
           MicroStrategy Inc.
         </a>
       </div>
-      <div className="column is-5 has-text-centered">
-        Software Engineer, Principal
-      </div>
+      <div className="column is-5 has-text-centered">Software Engineer, Principal</div>
       <div className="column is-3 has-text-right">01/2016 - present</div>
     </div>
     <div className="content">
       <ul>
         <li>
-          <div className="heading">Hackathons</div>
-          <p>I won the most recent three MicroStrategy Hackathons in a row.</p>
+          <div className="heading"><img className="icon is-small" src={OrangeLogo} alt="" /><span>Orange</span></div>
+          <p>
+            Developed the MicroStrategy Orange Outlook add-in, providing secure email communication through digital
+            signatures and verification. Built the Orange admin portal to manage user DIDs (decentralized identifiers).
+            As a hackathon project, I created a website and a browser extension for signing and verifying arbitrary
+            files and text messages. Demonstrated expertise in full-stack web development and a strong ability to
+            deliver innovative solutions.
+          </p>
+        </li>
+        <li>
+          <div className="heading"><img className="icon is-small" src={LightingLogo} alt="" /><span>Lighting Wallet</span></div>
+          <p>
+            Spearheaded the front-end development of the MicroStrategy Lightning Wallet, a digital wallet solution on
+            the Bitcoin{" "}
+            <a target="_blank" rel="noopener noreferrer" href="https://lightning.network/">
+              Lightning Network
+            </a>
+            . Designed and implemented cross-platform user interfaces, including a browser extension, a website, and
+            mobile applications. Created robust administrative and provisioning tools to support the Lightning Network
+            ecosystem.
+          </p>
         </li>
         <li>
           <div className="heading">
@@ -42,15 +53,10 @@ export const Experiences = () => (
             >
               Hyper Intelligence
             </a>
-            <span>
-              {' '}
-              - inject analytics and intelligence into every user&apos;s
-              experience.
-            </span>
+            <span> - inject analytics and intelligence into every user&apos;s experience.</span>
           </div>
           <p>
-            I took on the Engineering Manager role in the Hyper division. I led
-            the designs of HypeIntelligence for&nbsp;
+            I led the designs of HypeIntelligence for&nbsp;
             <a
               target="_blank"
               rel="noopener noreferrer"
@@ -74,10 +80,8 @@ export const Experiences = () => (
             >
               Desktop
             </a>
-            , built development infrastructures including CI/CD pipelines, i18n
-            workflows, and testing frameworks. I initiated software design
-            discussions and proposed a loosely coupled application architecture
-            named&nbsp;
+            , built development infrastructures including CI/CD pipelines, i18n workflows, and testing frameworks. I
+            initiated discussions on software design and introduced a flexible application architecture called&nbsp;
             <a
               target="_blank"
               rel="noopener noreferrer"
@@ -85,7 +89,7 @@ export const Experiences = () => (
             >
               Micro Module Services
             </a>
-            , it was recognized by the Chief Architect at MicroStrategy.
+            , which was praised by then Chief Architect.
           </p>
         </li>
         <li>
@@ -101,31 +105,22 @@ export const Experiences = () => (
             <span> - the unified tools for MicroStrategy Products</span>
           </div>
           <p>
-            As the technical lead, I participated in the design and development
-            of the Workstation macOS application. Solely implemented a number of
-            features including the R and Python integration and authentication.
-            With Jenkins, I built the CI/CD pipelines and i18n workflows that
-            saved 40% of the time for developers. In my spare time, I created a
-            web application named &quot;Lost In Translation&quot; that made
-            localization much easier, it was applied to most projects in
-            MicroStrategy. I also interviewed and mentored associate engineers.
+            As a Workstation macOS app developer, I independently developed key features like authentication and the R
+            and Python integration. I created the CI/CD pipelines on Jenkins, and designed the i18n workflows that saved
+            40% of the time for developers. Outside of regular work, I built a website called "Lost In Translation" to
+            simplify translating the app into different languages, many MicroStrategy projects used this tool. I also
+            hired and trained new engineers.
           </p>
         </li>
       </ul>
     </div>
     <div className="columns is-2">
       <div className="column is-4">
-        <a
-          target="_blank"
-          rel="noopener noreferrer"
-          href="https://www.microstrategy.com/cn"
-        >
+        <a target="_blank" rel="noopener noreferrer" href="https://www.microstrategy.com/cn">
           MicroStrategy China Technology Center Ltd.
         </a>
       </div>
-      <div className="column is-5 has-text-centered">
-        Principal Software Engineer II
-      </div>
+      <div className="column is-5 has-text-centered">Principal Software Engineer II</div>
       <div className="column is-3 has-text-right">07/2008 - 01/2016</div>
     </div>
     <div className="content">
@@ -144,38 +139,17 @@ export const Experiences = () => (
             </a>
           </div>
           <p>
-            Emma is a friendly marketplace like Facebook Marketplace, as a
-            Python developer I integrated both Facebook and PayPal APIs for
-            Emma. Usher is a secure enterprise identity management system, with
-            PHP and YII framework I led three engineers designed the
-            Usher&apos;s administration website - Network Manager, solely
-            implemented the logging module that supports keyword filtering,
-            created a client script minifier and a file uploading plugin for YII
-            in spare time that made the developers&apos; life much easier.
-          </p>
-        </li>
-        <li>
-          <div className="heading">
-            Quote System 2.0, Purchase Order, and Expense System
-          </div>
-          <p>
-            As a developer, I replaced traditional ASP.NET form with AJAX in the
-            Quote System 2.0 and saved 50% of page navigation; implemented the
-            Purchase Order website, the email module I added 10 years ago
-            supports templates and interceptions, it&apos;s so flexible thus
-            still actively used today; I also added Web Services for the Expense
-            System with WCF for mobile applications within a week.
+            I integrated Facebook and PayPal APIs into Emma, a marketplace similar to Facebook Marketplace. Leading a
+            team of three engineers, I designed Usher's Network Manager, a secure enterprise identity management system
+            built with PHP and Yii. I independently developed the logging module with keyword filtering capabilities,
+            created a client script minifier and a file uploading plugin for Yii to enhance developer efficiency.
           </p>
         </li>
       </ul>
     </div>
     <div className="columns is-2">
       <div className="column is-4">
-        <a
-          target="_blank"
-          rel="noopener noreferrer"
-          href="https://beyondsoft.com"
-        >
+        <a target="_blank" rel="noopener noreferrer" href="https://beyondsoft.com">
           Beyondsoft Consulting Inc.
         </a>
       </div>
@@ -183,12 +157,9 @@ export const Experiences = () => (
       <div className="column is-3 has-text-right">05/2007 - 08/2008</div>
     </div>
     <div className="content">
-      As an SDET, designed, and implemented test cases for automated functional
-      testing of Visual Studio IDE and both C# and VB.net programming languages.
-      I wrote a number of SQL scripts for the MadDog database, the scripts
-      aggregated the real-time status of test runs and generated well-formatted
-      reports. The scripts saved 20% of my time and were recognized and shared
-      in the team.
+      Created automated tests to ensure Visual Studio IDE and C#/.NET programming languages function correctly. I wrote
+      SQL scripts to combine real-time test results into clear reports, saving me 20% of time. These scripts were
+      well-received and shared by the team.
     </div>
   </section>
 );
