@@ -1,16 +1,11 @@
-import { download, fs } from 'node-build-tools';
-import path from 'path';
+import { download, fs } from "node-build-tools";
+import { join } from "node:path";
+import type { Config } from "./Config";
+import type { Profile } from "./Profile";
+import { ShieldFormat } from "./ShieldFormat";
+import { ShieldStyle } from "./ShieldStyle";
 
-import { Config } from './Config';
-import { Profile } from './Profile';
-import { ShieldFormat } from './ShieldFormat';
-import { ShieldStyle } from './ShieldStyle';
-
-export const getShieldImageFileName = (
-  profile: Profile,
-  style: string,
-  format: ShieldFormat
-) => {
+export const getShieldImageFileName = (profile: Profile, style: string, format: ShieldFormat) => {
   if (format === ShieldFormat.iconOnly) {
     return `${profile.label.toLowerCase()}@4x.png`;
   }
@@ -30,15 +25,12 @@ export const downloadShields = async (config: Config, dir: string) => {
       console.group(profile.label);
 
       for (const format of Object.values(ShieldFormat)) {
-        const leftText = format === ShieldFormat.labelled ? profile.label : '';
-        const rightText =
-          format === ShieldFormat.iconOnly ? '' : profile.username;
+        const leftText = format === ShieldFormat.labelled ? profile.label : "";
+        const rightText = format === ShieldFormat.iconOnly ? "" : profile.username;
         const rightBackgroundColor =
-          format === ShieldFormat.labelled
-            ? profile.rightBackgroundColor
-            : profile.leftBackgroundColor;
+          format === ShieldFormat.labelled ? profile.rightBackgroundColor : profile.leftBackgroundColor;
         const fileName = getShieldImageFileName(profile, style, format);
-        const saveAs = path.join(dir, fileName);
+        const saveAs = join(dir, fileName);
         const url = `https://img.shields.io/badge/${leftText}-${rightText}-${rightBackgroundColor}.svg?style=${style}&logo=${profile.iconName}&logoColor=${profile.iconColor}&labelColor=${profile.leftBackgroundColor}`;
 
         console.log(`${format}: ${url}`);
