@@ -1,11 +1,10 @@
-import fs from 'fs';
-
-import { Config } from './Config';
-import { getShieldImageFileName } from './downloadShields';
-import { placeholder } from './placeholder';
+import { readFileSync, writeFileSync } from "node:fs";
+import type { Config } from "./Config";
+import { getShieldImageFileName } from "./downloadShields";
+import { placeholder } from "./placeholder";
 
 const makeShields = ({ profiles, style, shieldFormat }: Config) =>
-  profiles.map((profile) => {
+  profiles.map(profile => {
     const shieldName = getShieldImageFileName(profile, style, shieldFormat);
     const src = `https://raw.githubusercontent.com/yanli0303/yanli0303/master/assets/${shieldName}`;
     const img = `<img alt="Yan Li - ${profile.label}" src="${src}" width="24" height="24" />`;
@@ -13,12 +12,7 @@ const makeShields = ({ profiles, style, shieldFormat }: Config) =>
     return `[${img}](${homepage})`;
   });
 
-const updateSection = (
-  lines: string[],
-  begin: string,
-  end: string,
-  replaceWith: string[]
-) => {
+const updateSection = (lines: string[], begin: string, end: string, replaceWith: string[]) => {
   let first = 0;
   let numberOfLines = -1;
   for (let i = 0; i < lines.length; i += 1) {
@@ -47,28 +41,23 @@ const updateSection = (
 };
 
 const updateShields = (lines: string[], config: Config) =>
-  updateSection(
-    lines,
-    '<!--SHIELDS_BEGIN-->',
-    '<!--SHIELDS_END-->',
-    makeShields(config)
-  );
+  updateSection(lines, "<!--SHIELDS_BEGIN-->", "<!--SHIELDS_END-->", makeShields(config));
 
 const updateGetInTouch = (lines: string[], config: Config) =>
   updateSection(
     lines,
-    '<!--GET_IN_TOUCH_BEGIN-->',
-    '<!--GET_IN_TOUCH_END-->',
-    config.profiles.map((p) => {
+    "<!--GET_IN_TOUCH_BEGIN-->",
+    "<!--GET_IN_TOUCH_END-->",
+    config.profiles.map(p => {
       const homepage = placeholder(p.urlTemplate, p);
-      const linkText = p.label === 'Email' ? p.username : homepage;
+      const linkText = p.label === "Email" ? p.username : homepage;
       return `- ${p.label}: [${linkText}](${homepage})`;
-    })
+    }),
   );
 
 const updateLastUpdated = (lines: string[]) => {
-  const lastUpdatedPrefix = 'Updated on: ';
-  return lines.map((line) => {
+  const lastUpdatedPrefix = "Updated on: ";
+  return lines.map(line => {
     if (line.startsWith(lastUpdatedPrefix)) {
       const lastUpdated = new Date().toString();
       return `${lastUpdatedPrefix}${lastUpdated}`;
@@ -78,9 +67,9 @@ const updateLastUpdated = (lines: string[]) => {
 };
 
 export const updateReadme = (readme: string, config: Config) => {
-  let contents = fs.readFileSync(readme, { encoding: 'utf-8' }).split(/\r?\n/);
+  let contents = readFileSync(readme, { encoding: "utf-8" }).split(/\r?\n/);
   contents = updateShields(contents, config);
   contents = updateGetInTouch(contents, config);
   contents = updateLastUpdated(contents);
-  fs.writeFileSync(readme, contents.join('\n'), { encoding: 'utf-8' });
+  writeFileSync(readme, contents.join("\n"), { encoding: "utf-8" });
 };
